Move main simulation routine into own file
