+++ src/commands.rs
@@ -50,7 +50,7 @@ pub fn handle_run_command(model_dir: &PathBuf) -> Result<()> {
     log::init(settings.log_level.as_deref()).context("Failed to initialize logging.")?;
     let model = Model::from_path(model_dir).context("Failed to load model.")?;
     info!("Model loaded successfully.");
-    crate::run(&model);
+    crate::simulation::run(&model);
     Ok(())
 }
