+++ src/lib.rs
@@ -1,4 +1,4 @@
-//! High level functionality for launching the simulation.
+//! Common functionality for MUSE 2.0.
 #![warn(missing_docs)]
 pub mod agent;
 pub mod commands;
@@ -10,20 +10,5 @@ pub mod model;
 pub mod process;
 pub mod region;
 pub mod settings;
+pub mod simulation;
 pub mod time_slice;
-
-use crate::model::Model;
-
-/// Run the simulation.
-///
-/// # Arguments:
-///
-/// * `model` - The model to run
-pub fn run(model: &Model) {
-    // TODO: Remove this once we're actually doing something with these values
-    println!("Commodities: {:?}", model.commodities);
-    println!("Regions: {:?}", model.regions);
-    println!("Processes: {:?}", model.processes);
-    println!("Time slices: {:?}", model.time_slice_info);
-    println!("Milestone years: {:?}", model.milestone_years);
-}
